Remove AssetLoader class
