+++ include/openspace/scene/assetmanager.h
@@ -25,10 +25,11 @@
 #ifndef __OPENSPACE_CORE___ASSETMANAGER___H__
 #define __OPENSPACE_CORE___ASSETMANAGER___H__
 
-#include <openspace/scene/assetloader.h>
+#include <openspace/util/synchronizationwatcher.h>
 #include <ghoul/lua/ghoul_lua.h>
 #include <ghoul/lua/luastate.h>
 #include <memory>
+#include <mutex>
 #include <unordered_set>
 #include <vector>
 
@@ -50,6 +51,7 @@ class SynchronizationWatcher;
 class AssetManager {
 public:
     AssetManager(ghoul::lua::LuaState* state, std::string assetRootDirectory);
+    ~AssetManager();
 
     void initialize();
     void deinitialize();
@@ -62,14 +64,78 @@ public:
     bool update();
     scripting::LuaLibrary luaLibrary();
 
+    /**
+     * Load an asset
+     */
+    bool loader_loadAsset(Asset* asset);
+
+    /**
+     * Unload an asset
+     */
+    void loader_unloadAsset(Asset* asset);
+
+    /**
+     * Call the onInitialize function specified in the asset file
+     */
+    void loader_callOnInitialize(Asset* asset);
+
+    /**
+     * Call the onDeinitialize function specified in the asset file
+     */
+    void loader_callOnDeinitialize(Asset* asset);
+
 private:
+    /**
+     * Add the asset as a request of the root asset
+     */
+    std::shared_ptr<Asset> loader_add(const std::string& identifier);
+
+    /**
+     * Remove the asset as a request of the root asset
+     */
+    void loader_remove(const std::string& identifier);
+    /**
+     * Return the asset identified by the identifier, if the asset is tracked. Otherwise 
+     * return nullptr.
+     */
+    std::shared_ptr<Asset> loader_has(const std::string& name) const;
+
+    /// Return the root asset
+    const Asset& loader_rootAsset() const;
+
+    /// Return the root asset
+    Asset& loader_rootAsset();
+
+    void loader_setUpAssetLuaTable(Asset* asset);
+    void loader_tearDownAssetLuaTable(Asset* asset);
+
+    std::shared_ptr<Asset> loader_asset(const std::string& name);
+    std::filesystem::path loader_currentDirectory() const;
+
+    void loader_setCurrentAsset(Asset* asset);
+
+
     std::unordered_set<std::string> _assetAddQueue;
     std::unordered_set<std::string> _assetRemoveQueue;
     std::mutex _pendingInitializationsMutex;
     std::vector<std::shared_ptr<Asset>> _pendingInitializations;
 
     SynchronizationWatcher _synchronizationWatcher;
-    AssetLoader _assetLoader;
+    //AssetLoader _assetLoader;
+
+
+    // Member variables
+    std::shared_ptr<Asset> _rootAsset;
+    Asset* _currentAsset = nullptr;
+    std::unordered_map<std::string, std::weak_ptr<Asset>> _trackedAssets;
+    std::string _assetRootDirectory;
+    ghoul::lua::LuaState* _luaState;
+
+    // References to Lua values
+    std::unordered_map<Asset*, std::vector<int>> _onInitializeFunctionRefs;
+    std::unordered_map<Asset*, std::vector<int>> _onDeinitializeFunctionRefs;
+
+    int _assetsTableRef = 0;
 };
 
 } // namespace openspace
